+++ scripts/recreate_darwin_dev.sql
@@ -237,7 +237,7 @@ CREATE TABLE priority_card_order (
 
 CREATE TABLE map_routes (
     id              INT             NOT NULL PRIMARY KEY AUTO_INCREMENT,
-    route_id        INT             NOT NULL,
+    route_id        BIGINT          NOT NULL,
     name            VARCHAR(256)    NOT NULL,
     creator_fk      VARCHAR(64)     NOT NULL,
     create_ts       TIMESTAMP       NULL DEFAULT CURRENT_TIMESTAMP,
@@ -250,7 +250,7 @@ CREATE TABLE map_routes (
 
 CREATE TABLE map_runs (
     id              INT             NOT NULL PRIMARY KEY AUTO_INCREMENT,
-    run_id          INT             NOT NULL,
+    run_id          BIGINT          NOT NULL,
     map_route_fk    INT             NULL,
     activity_id     INT             NOT NULL,
     activity_name   VARCHAR(16)     NOT NULL,
